# docs/api/swagger.yaml
definitions:
  APIErrorResponse:
    properties:
      code:
        type: integer
      details: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_flotio-dev_core-api_internal_modules_user_model.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: string
      refresh_token:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_modules_user_model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_modules_user_model.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_modules_user_model.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_modules_user_model.StatusResponse:
    properties:
      status:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_modules_user_model.UpdateUserRequest:
    properties:
      email:
        type: string
      github_id:
        type: string
      github_username:
        type: string
      username:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_modules_user_model.UserResponse:
    properties:
      created:
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
      username:
        example: johndoe
        type: string
    type: object
...
info:
  contact:
    email: support@flotio.dev
    name: Flotio Team
  description: Flotio Core API — authentication, environment/keystore/Google Play
    credentials assets, projects, build pipeline, releases and GitHub integration
    for the Flotio mobile CI platform.
  license:
    name: Proprietary
  title: Flotio Core API
  version: 1.0.0
paths:
  /auth/@me:
    get:
      description: Return information about the authenticated user
      operationId: MeGetHandler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Update email and/or username of the authenticated user
      operationId: MePutHandler
      parameters:
      - description: User update payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return access & refresh tokens
      operationId: LoginHandler
      parameters:
      - description: Login payload
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke a refresh token and logout the user
      operationId: LogoutHandler
      parameters:
      - description: Refresh token payload
        in: body
        name: logout
        required: true
        schema:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a valid refresh token (rotation
        enabled)
      operationId: RefreshTokenHandler
      parameters:
      - description: Refresh token payload
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and return access & refresh tokens
      operationId: RegisterHandler
      parameters:
      - description: Register payload
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_flotio-dev_core-api_internal_modules_user_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"